+++ python/attoworld/file/__init__.py
@@ -1,11 +1,10 @@
 """
 This module will contain functions for loading the various data formats used across the labs.
 """
-
 from .file_io import load_waves_from_matfile, load_waveform_from_text, load_spectrum_from_text, read_Trebino_FROG_matrix, read_Trebino_FROG_speck, read_Trebino_FROG_data
-from .interface_simulations import LunaResult
-from .profileAndIntensity import profile_analysis
+from .LunaResult import LunaResult
 from .dataclasses import Waveform, ComplexSpectrum, IntensitySpectrum, ComplexEnvelope, FrogData
+
 __all__ = ["load_waves_from_matfile",
     "load_waveform_from_text",
     "load_spectrum_from_text",
@@ -14,7 +13,6 @@ __all__ = ["load_waves_from_matfile",
     "read_Trebino_FROG_speck",
     "read_Trebino_FROG_data",
     "LunaResult",
-    "profile_analysis",
     "Waveform",
     "ComplexSpectrum",
     "IntensitySpectrum",
