+++ python/attoworld/personal/__init__.py
@@ -5,3 +5,10 @@ of data processing files that utilize functions that might not otherwise be of g
 
 from . import nick
 from . import vlad
+from . import marco
+
+__all__ = [
+    'nick',
+    'vlad',
+    'marco'
+]
