+++ python/attoworld/personal/nick/__init__.py
@@ -3,5 +3,5 @@ This file contains functions primarily used by Nick Karpowicz.
 """
 
 from .nick import dark_plot
-
-__all__ = ["dark_plot"]
+from .impulse_response import get_effective_response
+__all__ = ["dark_plot", "get_effective_response"]
